ClamAV implementation

+++ docker-ops.sh
@@ -11,8 +11,7 @@ LOCAL_IMAGE="local/${IMAGE_NAME}:${IMAGE_VERSION}"
 DOCKER_REPOS=("htec" "public.ecr.aws/htec")
 
 build() {
-  docker build -t "${LOCAL_IMAGE}" --build-arg "VERSION=${IMAGE_VERSION}" \
-    -f "${IMAGE_NAME}/Dockerfile" .
+  docker build -t "${LOCAL_IMAGE}" --build-arg "VERSION=${IMAGE_VERSION}" "${IMAGE_NAME}/"
 }
 
 push() {
@@ -27,8 +26,13 @@ push() {
 }
 
 cleanup() {
-  docker rmi "${LOCAL_IMAGE}"
+  if [ "${SKIP_CLEANUP}" == "true" ]; then
+      echo "Skipping cleanup"
+  else
+      docker rmi "${LOCAL_IMAGE}"
+  fi;
+
 }
 
-build && push && cleanup
+build && push  && cleanup
